+++ automation/src/epilot/models/shared/apisubmissiontrigger.py
@@ -3,8 +3,9 @@
 from __future__ import annotations
 import dataclasses
 from dataclasses_json import Undefined, dataclass_json
+from enum import Enum
 from epilot import utils
-from typing import Final, Optional
+from typing import Optional
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -15,12 +16,15 @@ class APISubmissionTriggerConfiguration:
     
 
 
+class APISubmissionTriggerType(str, Enum):
+    API_SUBMISSION = 'api_submission'
+
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 
 @dataclasses.dataclass
 class APISubmissionTrigger:
     configuration: APISubmissionTriggerConfiguration = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('configuration') }})
-    TYPE: Final[str] = dataclasses.field(default='api_submission', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: APISubmissionTriggerType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})