+++ automation/src/epilot/models/shared/cartcheckoutactionconfig.py
@@ -4,8 +4,12 @@ from __future__ import annotations
 import dataclasses
 from ..shared import cartcheckoutconfig as shared_cartcheckoutconfig
 from dataclasses_json import Undefined, dataclass_json
+from enum import Enum
 from epilot import utils
-from typing import Final, Optional
+from typing import Optional
+
+class CartCheckoutActionConfigType(str, Enum):
+    CART_CHECKOUT = 'cart-checkout'
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -21,6 +25,6 @@ class CartCheckoutActionConfig:
     flow_action_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('flow_action_id'), 'exclude': lambda f: f is None }})
     id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('id'), 'exclude': lambda f: f is None }})
     name: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('name'), 'exclude': lambda f: f is None }})
-    TYPE: Final[Optional[str]] = dataclasses.field(default='cart-checkout', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})
+    type: Optional[CartCheckoutActionConfigType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})