+++ automation/src/epilot/models/shared/createdocumentactionconfig.py
@@ -4,8 +4,12 @@ from __future__ import annotations
 import dataclasses
 from ..shared import createdocumentconfig as shared_createdocumentconfig
 from dataclasses_json import Undefined, dataclass_json
+from enum import Enum
 from epilot import utils
-from typing import Final, Optional
+from typing import Optional
+
+class CreateDocumentActionConfigType(str, Enum):
+    CREATE_DOCUMENT = 'create-document'
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -20,6 +24,6 @@ class CreateDocumentActionConfig:
     flow_action_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('flow_action_id'), 'exclude': lambda f: f is None }})
     id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('id'), 'exclude': lambda f: f is None }})
     name: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('name'), 'exclude': lambda f: f is None }})
-    TYPE: Final[Optional[str]] = dataclasses.field(default='create-document', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})
+    type: Optional[CreateDocumentActionConfigType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})