+++ automation/src/epilot/models/shared/entitymanualtrigger.py
@@ -3,8 +3,9 @@
 from __future__ import annotations
 import dataclasses
 from dataclasses_json import Undefined, dataclass_json
+from enum import Enum
 from epilot import utils
-from typing import Final, Optional
+from typing import Optional
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -16,12 +17,15 @@ class EntityManualTriggerConfiguration:
     
 
 
+class EntityManualTriggerType(str, Enum):
+    ENTITY_MANUAL = 'entity_manual'
+
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 
 @dataclasses.dataclass
 class EntityManualTrigger:
     configuration: EntityManualTriggerConfiguration = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('configuration') }})
-    TYPE: Final[str] = dataclasses.field(default='entity_manual', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: EntityManualTriggerType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})